more

+++ builtin/setenv/env_var_change.c
@@ -6,16 +6,16 @@
 /*   By: zraunio <zraunio@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/09/30 11:56:16 by zraunio           #+#    #+#             */
-/*   Updated: 2021/09/30 11:58:04 by zraunio          ###   ########.fr       */
+/*   Updated: 2021/10/07 10:57:53 by zraunio          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../includes/setenv.h"
+#include "../../includes/setenv.h"
 
 void	env_var_change(t_shell *data, char *var, char *val, int i)
 {
-	int	ii;
-	int	x;
+	size_t	ii;
+	size_t	x;
 
 	if (ft_strcmp(val, "NULL") == 0)
 	{
@@ -23,8 +23,8 @@ void	env_var_change(t_shell *data, char *var, char *val, int i)
 		return (remove_one_string_in_array(data->environ, i));
 	}
 	ft_memdel((void *)&data->environ[i]);
-	data->environ[i] = ft_strnew((size_t)ft_strlen(var)
-			+ (size_t)ft_strlen(val) + 1 + 1);
+	data->environ[i] = ft_strnew(ft_strlen(var)
+			+ ft_strlen(val) + 1 + 1);
 	ii = 0;
 	x = 0;
 	while (var[ii] != '\0')
